add sortedChat && rerender chatFrom

+++ src/helpers/UI/ContextMenu.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import React, { FC, useState } from 'react';
 import style from './Helpers.module.scss';
 import axios from 'axios';
 import { useDispatch } from 'react-redux';
@@ -6,6 +6,7 @@ import { useTypedSelector } from '../../hooks/useTypedSelector';
 import { urlChats } from '../../url/url';
 import { modalErrorState, modalErrorText } from '../../store/reducers/modalErrorReducer';
 import { userOwner } from '../userOwner';
+
 
 
 interface ContextMenuProps {
@@ -17,6 +18,8 @@ interface ContextMenuProps {
 
 const ContextMenu: FC<ContextMenuProps> = ({ setOpenContextMenu, clientX, clientY, onRemove }) => {
 
+
+  
   const dispatch = useDispatch();
 
   const openModalErrorNoMessage = () => {
@@ -29,26 +32,26 @@ const ContextMenu: FC<ContextMenuProps> = ({ setOpenContextMenu, clientX, client
     dispatch(modalErrorText('Choose your message to delete'));
   };
 
-
-
   const { chosenMessageId: id } = useTypedSelector(state => state.chosenMessage);
   const { chosenMessageUserFrom: userFrom } = useTypedSelector(state => state.chosenMessage);
-
-
+  const [userFromState, setUserFromState] = useState(userFrom);
   const removeMessage = (event: React.MouseEvent<HTMLDivElement>) => {
-    if (userFrom == userOwner || userFrom == null) {
+    setUserFromState(userFrom);
+    console.log(userFromState);
+    if (userFrom == userOwner || userFrom == null && id) {
       setOpenContextMenu(false);
       event.stopPropagation();
       axios.delete(`${urlChats}/${id}`).then(() => {
         onRemove();
       }).catch(openModalErrorNoMessage);
-
     }
     else {
       setOpenContextMenu(false);
       event.stopPropagation();
       openModalErrorNotYourMessage();
     }
+    setUserFromState(0);
+    console.log(userFromState);
   };
